Validation of group device linking and delinking.

+++ src/main/java/org/traccar/api/resource/GroupPermissionResource.java
@@ -8,7 +8,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.traccar.api.BaseResource;
 import org.traccar.api.security.PermissionsService;
+import org.traccar.model.Group;
+import org.traccar.model.GroupDevice;
 import org.traccar.model.LinkType;
+import org.traccar.model.User;
+import org.traccar.storage.StorageException;
+import org.traccar.storage.query.Columns;
+import org.traccar.storage.query.Condition;
+import org.traccar.storage.query.Request;
 
 
 @Path("group/permissions")
@@ -24,16 +31,37 @@ public class GroupPermissionResource extends BaseResource {
     @POST
     public Response linkGroupDevice(@QueryParam("groupId") long groupId, @QueryParam("deviceId") long deviceId) throws Exception{
         //LOGGER.info("Received POST request -> groupId: {}, deviceId: {}", groupId, deviceId);
-        permissionsService.link(LinkType.GROUP_DEVICE, groupId, deviceId);
-        //return Response.ok().build();
-        return Response.ok("{\"status\":\"Linked successfully\"}").build();
+        if(!validateGroupDeviceLink(groupId,deviceId)){
+            permissionsService.link(LinkType.GROUP_DEVICE, groupId, deviceId);
+            return Response.ok("{\"status\":\"Linked successfully\"}").build();
+        }else{
+            return Response.ok("{\"status\":\"Device already linked to the group.\"}").build();
+        }
+
     }
 
     @DELETE
     public Response unlinkGroupDevice(@QueryParam("groupId") long groupId, @QueryParam("deviceId") long deviceId) throws Exception{
-        permissionsService.unlink(LinkType.GROUP_DEVICE, groupId, deviceId);
-        //return Response.ok().build();
-        return Response.ok("{\"status\":\"Unlinked successfully.\"}").build();
+        if(validateGroupDeviceLink(groupId,deviceId)) {
+            permissionsService.unlink(LinkType.GROUP_DEVICE, groupId, deviceId);
+            return Response.ok("{\"status\":\"Unlinked successfully.\"}").build();
+        }else{
+            return Response.ok("{\"status\":\"The group-device link doesnot exist.\"}").build();
+        }
     }
+
+    public boolean validateGroupDeviceLink(long groupId, long deviceId) throws StorageException {
+        // Query the database for a record matching both groupId and deviceId
+        GroupDevice link = storage.getObject(GroupDevice.class, new Request(
+                new Columns.All(),
+                new Condition.And(
+                        new Condition.Equals("groupid", groupId),
+                        new Condition.Equals("deviceid", deviceId)
+                )));
+
+        // If the record exists, return true; otherwise, false
+        return link != null;
+    }
+
 
 }
